groundwork for new accounts system stored in the race store
