+++ cmd/server-manager/main.go
@@ -17,17 +17,21 @@ import (
 
 var debug = os.Getenv("DEBUG") == "true"
 
+var defaultAddress = "0.0.0.0:8772"
+
 func main() {
 	config, err := servermanager.ReadConfig("config.yml")
 
 	if err != nil {
-		logrus.Fatalf("could not open config file, err: %s", err)
+		ServeHTTPWithError(defaultAddress, "Read configuration file (config.yml)", err)
+		return
 	}
 
 	store, err := config.Store.BuildStore()
 
 	if err != nil {
-		logrus.Fatalf("could not open store, err: %s", err)
+		ServeHTTPWithError(defaultAddress, "Open server manager storage (bolt or json)", err)
+		return
 	}
 
 	servermanager.SetupRaceManager(store)
@@ -36,7 +40,8 @@ func main() {
 	err = servermanager.InstallAssettoCorsaServer(config.Steam.Username, config.Steam.Password, config.Steam.ForceUpdate)
 
 	if err != nil {
-		logrus.Fatalf("could not install assetto corsa server, err: %s", err)
+		ServeHTTPWithError(defaultAddress, "Install assetto corsa server with steamcmd. Likely you do not have steamcmd installed correctly.", err)
+		return
 	}
 
 	var templateLoader servermanager.TemplateLoader
@@ -57,7 +62,8 @@ func main() {
 	servermanager.ViewRenderer, err = servermanager.NewRenderer(templateLoader, debug)
 
 	if err != nil {
-		logrus.Fatalf("could not initialise view renderer, err: %s", err)
+		ServeHTTPWithError(defaultAddress, "Initialise view renderer (internal error)", err)
+		return
 	}
 
 	go servermanager.LoopRaces()
@@ -76,7 +82,8 @@ func main() {
 	listener, err := net.Listen("tcp", config.HTTP.Hostname)
 
 	if err != nil {
-		logrus.Fatalf("could not listen on address: %s, err: %s", config.HTTP.Hostname, err)
+		ServeHTTPWithError(defaultAddress, "Listen on hostname " + config.HTTP.Hostname + ". Likely the port has already been taken by another application", err)
+		return
 	}
 
 	logrus.Infof("starting assetto server manager on: %s", config.HTTP.Hostname)
