+++ cmd/server-manager/views/layout/base.html
@@ -75,10 +75,11 @@
                             Server
                         </a>
                         <div class="dropdown-menu" aria-labelledby="navBarServerDropdown">
-                            <a class="dropdown-item" href="/logs">Logs</a>
                             {{ if AdminAccess }}
                                 <a class="dropdown-item" href="/server-options">Options</a>
+                                <a class="dropdown-item" href="/accounts">Accounts</a>
                             {{ end }}
+                            <a class="dropdown-item" href="/logs">Logs</a>
                         </div>
                     </li>
                 </ul>
