+++ config_ini.go
@@ -149,6 +149,9 @@ type GlobalServerConfig struct {
 
 	FreeUDPPluginLocalPort int    `ini:"-" show:"-"`
 	FreeUDPPluginAddress   string `ini:"-" show:"-"`
+
+	// ServerFirstLaunch is used to prompt the user with configuration they need to do
+	ServerFirstLaunch bool `ini:"-" show:"-"`
 }
 
 type CurrentRaceConfig struct {
